MQTTService/IoT/requirements.txt
requests
pymongo
paho-mqtt